Add QSettings(Scope...) constructor to QSettings
Because of the system specific folder/filenames (organization
name vs domain) it was not possible to explicitly access the
settings with SystemScope without using #ifdef, as it is done
by Qt internally. The new constructor uses the default name
while creating an instance with a scope.

[ChangeLog][QtCore][QSettings] Added QSettings(Scope...)
constructor to avoid using #ifdef in Qt applications.

Change-Id: I81016430a1d18a382bfdd1e1cf32de367f98d7aa
Reviewed-by: Oswald Buddenhagen <oswald.buddenhagen@qt.io>
Reviewed-by: Edward Welbourne <edward.welbourne@qt.io>