Supply locale to QCollatorPrivate via its constructor
Both places that new'd it were then setting its locale themselves;
they might as well do it tidily by passing it to the constructor.
There's also no need to cleanup() in the constructor; every back-end
is a no-op when collator has its initial value.

Tidied up the class declaration in the process:
 * moved {con,de}structor to be first methods
 * comment on the two methods back-ends provide

Change-Id: I041669637935e68141e002156552af8b475ba36e
Reviewed-by: Thiago Macieira <thiago.macieira@intel.com>