Windows: Don't output a warning when sHGetKnownFolderPath fails
In the rare case where the known locations for the standard paths are
not known (such as when an application is used without a user logged
in), it will output a warning to indicate this. In the case of the
GenericConfigLocation, this can mean that it will hang due to the fact
that QLoggingCategory is looking for that location too before it can
output anything.

Therefore, the warning output is removed so that if this part fails it
doesn't cause it to hang as a result.

Change-Id: I4f189361899bd1f868292f30c09fbe50982d2288
Reviewed-by: Kai Koehne <kai.koehne@qt.io>
(cherry picked from commit bebae3737624a54f6f8062f1cbf32179fb43df7a)